Move TestList into Test module
