+++ frontend/src/pages/admin/index.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import AppFrame from '../../components/AppFrame';
 import AppContent from '../../components/AppContent';
-import TestList from '../../components/TestList';
+import TestList from '../../components/Test/List';
 
 const PageAdmin = () => (
   <AppFrame>
